% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-points.R
\name{as_points}
\alias{as_points}
\title{Convert to point}
\source{
\url{https://twitter.com/andrewheiss/status/1095450321784950785}
}
\usage{
as_points(x, input_unit = "mm")
}
\arguments{
\item{x}{Input.}

\item{input_unit}{Unit of \code{x}, default is millimeters which is default in \code{ggplot2::geom_text}, among other functions.}
}
\value{
Input units converted to points.
}
\description{
Convert input to actual points. Useful for setting size of text elements in \code{ggplot2}.
}
