% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{clipboard}
\alias{clipboard}
\title{Copy object to clipboard}
\usage{
clipboard(df)
}
\arguments{
\item{df}{input object, can be piped through \code{test}}
}
\value{
None
}
\description{
Copy object to clipboard
}
