% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coding_mutations}
\alias{coding_mutations}
\title{Coding  mutations}
\format{
Vector of "Variant_Classification" entries in a MAF file that qualify as coding mutations.
}
\usage{
coding_mutations
}
\description{
Coding  mutations
}
\keyword{datasets}
