% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{colorspace64}
\alias{colorspace64}
\title{Custom color palettes}
\format{An object of class \code{character} of length 104.}
\usage{
colorspace64
}
\arguments{
\item{n}{number of colors to return}
}
\value{
A vector colors the length of \code{n}
}
\description{
Palette \code{many_colors} contains 103 reasonably distinct colors
}
\details{
Palette \code{arsenal_colors} contains 3 distinct colors
}
\keyword{datasets}
