% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncokb_annotate_maf.R
\docType{data}
\name{consequence_map}
\alias{consequence_map}
\title{Annotate read MAF with OncoKB annotation}
\format{An object of class \code{character} of length 11.}
\source{
\url{oncokb.org}

\url{github.com/oncokb/oncokb-annotator}
}
\usage{
consequence_map
}
\arguments{
\item{maf}{Input MAF}

\item{cancer_types}{Data frame with samples mapped to cancer type for accurate levels of actionability}
}
\value{
Annotated MAF with columns indicating functionality of mutation and levels of actionability
}
\description{
Annotate read MAF with OncoKB annotation
}
\examples{

Sources of default hotspots below

}
\keyword{datasets}
