% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnames.R
\name{dnames}
\alias{dnames}
\title{Custom color palettes}
\usage{
dnames(name_vector)
}
\arguments{
\item{name_vector}{Vector of name.}
}
\value{
The duplicated values in \code{name_vector}.
}
\description{
Works like \code{duplicated} but returns the actual values.
}
