% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-vaf.R
\name{expected_vaf_germline}
\alias{expected_vaf_germline}
\alias{expected_vaf_somatic}
\title{Estimate somatic variant allele fraction (VAF)}
\usage{
expected_vaf_germline(purity, mcn, lcn)

expected_vaf_somatic(purity, mcn, lcn)
}
\arguments{
\item{maf}{Input MAF}

\item{hotspot}{Custom list of hotspots}
}
\value{
Object containing expected VAFs if the tumor is on the major (\code{mcn}) or minor allele (\code{lcn}), respectively.
}
\description{
Given copy-number and purity estimates, calculates the two expected alllele fractions of a mutation in a tumor.
Use \code{expected_vaf_germline} for germline mutations and \code{expected_vaf_somatic} for somatic mutations
}
