% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets-closeup.R
\name{facets_closeup}
\alias{facets_closeup}
\title{Generate gene-level Facets plot}
\usage{
facets_closeup(path, gene, chrom_range = NULL, plot = T,
  adjust_dlr = F)
}
\arguments{
\item{path}{Path to Rdata oject}

\item{gene}{One or multiple genes to plot}

\item{chrom_range}{Range of chromosomes}

\item{plot}{Plot or save as object}

\item{adjust_dlr}{Set dipLogR to 0, essentially normalizing the log ratio}
}
\value{
If \code{plot == T} plots cnlr, valor and icnem objects else returns all as objects
}
\description{
Generate gene-level Facets plot
}
\examples{
\donttest{facets_closeup('path_to_sample.Rdata', 'BRCA1', 17)}
}
