% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-themes.R
\docType{data}
\name{ggplot2_themes}
\alias{ggplot2_themes}
\alias{theme_bwmin}
\alias{theme_pub}
\title{\code{ggplot2} themes}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.

An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
theme_bwmin

theme_pub
}
\value{
None
}
\description{
\code{theme_bw} is a simple black-and white theme for interactive plotting.
\code{theme_pub} is a theme for producing publication-ready plots, with appropriate font and line sizing.
}
\examples{
library(ggplot2)
theme_set(theme_bwmin)
ggplot() + theme_pub

}
\keyword{datasets}
