% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_annotate_maf.R
\docType{data}
\name{hs_from_local}
\alias{hs_from_local}
\title{Annotate read MAF with hotspots}
\format{An object of class \code{character} of length 3.}
\source{
\url{cancerhotspots.org}

\url{3dhotspots.org}

\url{www.ncbi.nlm.nih.gov/pubmed/26619011}
}
\usage{
hs_from_local
}
\arguments{
\item{maf}{Input MAF}

\item{hotspot}{Custom list of hotspots}
}
\value{
Annotated MAF with columns \code{snv_hotspot}, \code{threeD_hotspot}, \code{indel_hotspot_type} and \code{Hotspot} indicating types of hotspots
}
\description{
Sources of default hotspots below
}
\keyword{datasets}
