% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if-na.R
\name{if_na}
\alias{if_na}
\title{Replace \code{NA}s}
\usage{
if_na(x, replacement)
}
\arguments{
\item{x}{What to replace}

\item{replacement}{What to replace it with}
}
\value{
Vector where \code{x} is replaced by \code{replacement}
}
\description{
Abbreviated \code{ifelse} replacement
}
