% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{many_colors}
\alias{many_colors}
\title{Custom color palettes}
\source{
\url{http://godsnotwheregodsnot.blogspot.ru/2012/09/color-distribution-methodology.html}
}
\usage{
many_colors(n)
}
\arguments{
\item{n}{number of colors to return.}
}
\value{
A vector colors the length of \code{n}
}
\description{
Palette \code{many_colors} contains 103 reasonably distinct colors.
Palette \code{arsenal_colors} contains 3 distinct colors.
}
