% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncokb-annotate-cna.R
\name{oncokb_annotate_cna}
\alias{oncokb_annotate_cna}
\title{Annotate OncoKB levels of evidence}
\source{
\url{oncokb.org}

\url{github.com/oncokb/oncokb-annotator}
}
\usage{
oncokb_annotate_cna(cnas, cancer_types = NULL)
}
\arguments{
\item{cnas}{Input CNAs}

\item{cancer_types}{Data frame with samples mapped to cancer type for accurate levels of actionability.}

\item{gene}{Gene.}

\item{cna_type}{Type of CNA, mapped from litterals and numeric encoding.}
}
\value{
Annotated CNAs with columns indicating functionality of mutation and levels of actionability.
}
\description{
Adds OncoKB oncogenicity and actionability annotation to list of copy-number alterations. See URLs below.
}
