% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-stratton.R
\name{plot_stratton}
\alias{plot_stratton}
\title{Stratton plot}
\usage{
plot_stratton(maf, sample_name = NULL, return_object = FALSE)
}
\arguments{
\item{maf}{Input MAF}

\item{sample_name}{Select sample or all present in MAF}
}
\value{
\code{ggplot2} plot
}
\description{
Stratton style plot of trinucleotide context of SNVs. Can plot one or more samples
}
