% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-survival.R
\name{plot_survival}
\alias{plot_survival}
\alias{parse_survfit}
\alias{plot_survfit}
\title{Plot Kaplan-Meier plot}
\usage{
parse_survfit(survfit_object)

plot_survfit(survfit_object, colors = NULL, legend_title = NULL,
  plot_title = NULL, xmax = NULL, parse_strata = FALSE,
  x_lab = NULL, y_lab = NULL, legend_pos = NULL, pval = TRUE,
  pval_pos = NULL, interval = NULL, env_theme = TRUE,
  table = FALSE)
}
\arguments{
\item{survfit_object}{Input \code{survfit} object}

\item{colors}{Colors for strata}

\item{legend_title}{Title of plot legend}

\item{plot_title}{Title of plot}

\item{xmax}{Limit for x axis}

\item{parse_strata}{Convert 0/1 notation to wildtype/mutated}

\item{x_lab}{Label x axis}

\item{y_lab}{Label y axis}

\item{legend_pos}{Figure legend position}

\item{pval}{Show p-value of model}

\item{pval_pos}{P-value position}

\item{interval}{Tick interval on x axis}

\item{env_theme}{Theme, if not current}

\item{table}{Show at-risk table beneath plot (experimental)}
}
\value{
\code{ggplot2}-based Kaplan-Meier plot
}
\description{
From a \code{survfit}, plot stratified Kaplan-Meier plot.
}
