% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal-link.R
\name{portal_link}
\alias{portal_link}
\title{cBioPortal link}
\source{
\url{https://cbioportal.mskcc.org}
}
\usage{
portal_link(samples)
}
\arguments{
\item{sample}{Vector of one or more samples to look up}
}
\value{
URL to sample(s) at cBioPortal
}
\description{
Creates URL based on listed samples to instance of MSK-IMPACT samples
}
