% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Read input file}
\usage{
read(x, ...)
}
\arguments{
\item{x}{Input file}

\item{...}{Any extra argument passed to `vroom::vroom`}
}
\value{
The result from `vroom::vroom`
}
\description{
Shorthand for reading files with `vroom::vroom` suppressing messages
}
