% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_germline.R
\name{read_germline}
\alias{read_germline}
\title{Read DMP distributed germline file}
\usage{
read_germline(filepath)
}
\arguments{
\item{path}{Excel file}
}
\value{
Data frame
}
\description{
Properly parse fields in DMP's germline distributed files
}
