% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_impact_repository_files.R
\name{read_impact_repository_files}
\alias{read_impact_repository_files}
\alias{read_impact_maf}
\alias{read_impact_cna}
\alias{read_impact_fusions}
\alias{read_impact_samples}
\alias{read_impact_patients}
\title{Helper functions to read files from MSK-IMPACT repository}
\usage{
read_impact_maf(filename = NULL, unfiltered = F, germline = F)

read_impact_cna(filename = NULL)

read_impact_fusions(filename = NULL)

read_impact_samples(filename = NULL)

read_impact_patients(filename = NULL)
}
\arguments{
\item{filename}{Override default filenames, assumes they exist in default location}

\item{unfiltered}{Read MAF containing non-signed out mutation calls}

\item{germline}{Retain germline calls in MAF}
}
\value{
Data object in tweaked cBioPortal format
}
\description{
Helper functions that reads the most commonly used filed from the MSK-IMPACT Mercurial repository.
}
