% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-maf.R
\name{read_maf}
\alias{read_maf}
\title{Read MAF file}
\usage{
read_maf(maf_path)
}
\arguments{
\item{maf_path}{Path to input file}
}
\value{
None
}
\description{
Simple function to read MAF file and a \code{t_var_freq} column
}
