% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_signatures.R
\docType{data}
\name{read_signatures}
\alias{read_signatures}
\alias{signatures_map}
\alias{read_signatures}
\title{Mutational signatures}
\format{An object of class \code{character} of length 30.}
\source{
\url{cancer.sanger.ac.uk/cosmic/signatures}
}
\usage{
signatures_map

read_signatures(input_file)
}
\arguments{
\item{input_file}{Path to output from \code{main.py} in \url{github.com/mskcc/mutation-signatures}}
}
\value{
Melted data frame with trivial names of signatures
}
\description{
Read output from mutational signatures decomposition
}
\keyword{datasets}
