% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_theme.R
\name{set_theme}
\alias{set_theme}
\title{Set \code{ggplot2} theme}
\usage{
set_theme(user_theme)
}
\arguments{
\item{user_theme}{Name of theme}
}
\description{
Basic function to set \code{ggplot2} theme, also setting default fonts for all text objects
}
