% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setlist.R
\name{setlist}
\alias{setlist}
\title{Set operations}
\usage{
setlist(a, b, output = F)
}
\arguments{
\item{a}{First input vector of items}

\item{b}{First input vector of items}

\item{output}{Return object with intersect/union of input sets}
}
\value{
A list object, if \code{output == TRUE}
}
\description{
Returns basic set operations on two lists
}
