% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_barplot.R
\name{signature_barplot}
\alias{signature_barplot}
\title{Signature barplot}
\usage{
signature_barplot(sign, col_vec = NULL, sample_order = NULL,
  sign_order = NULL, labels = TRUE)
}
\arguments{
\item{sign}{Signature data, as from \code{read_signatures}}

\item{col_vec}{Override plot colors}

\item{sample_order}{Override ordering of samples}

\item{sign_order}{Override ordering of signatures}

\item{labels}{Override labels}
}
\value{
A list object, if \code{output == TRUE}
}
\description{
Simple barplot of mutational signatues across samples
}
