% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort-oncoprint.R
\name{sort_oncoprint}
\alias{sort_oncoprint}
\title{Sort oncoprint}
\usage{
sort_oncoprint(M)
}
\arguments{
\item{M}{Input matrix}
}
\value{
A \code{list} object containing order of genes and samples.
}
\description{
Sort a gene X sample matrix with binary fields according to oncoprint method
}
