% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/test-maf.R
\docType{data}
\name{test_maf}
\alias{test_maf}
\title{Test MAF}
\format{Mutation annotation format (MAF) with standard columns.}
\source{
\url{portal.gdc.cancer.gov}

\url{https://github.com/oncokb/oncokb-annotator}
}
\usage{
test_maf

test_maf
}
\description{
Test MAF

Example MAF for testing
}
\keyword{datasets}
