% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bwmin.R
\docType{data}
\name{theme_bwmin}
\alias{theme_bwmin}
\title{\code{ggplot2} theme}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 60.}
\usage{
theme_bwmin
}
\value{
None
}
\description{
Simple black-and-white theme for \code{ggplot2}
}
\examples{
theme_set(theme_bwmin)
ggplot() + theme_bwmin

}
\keyword{datasets}
