% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pub.R
\docType{data}
\name{theme_pub}
\alias{theme_pub}
\title{\code{ggplot2} theme}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 65.}
\usage{
theme_pub
}
\value{
None
}
\description{
Simple almost-publication-ready theme for \code{ggplot2}
}
\examples{
theme_set(theme_pub)
ggplot() + theme_pub
}
\keyword{datasets}
