% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{truncating_mutations}
\alias{truncating_mutations}
\title{Truncating mutations}
\format{Vector of "Variant_Classification" entries in a MAF file that qualify as protein-truncating mutations.}
\usage{
truncating_mutations
}
\description{
Truncating mutations
}
\keyword{datasets}
