% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_out.R
\name{write_out}
\alias{write_out}
\title{Write function}
\usage{
write_out(x, file, col.names = T)
}
\arguments{
\item{x}{Object to write}

\item{file}{File to write to}

\item{col.names}{Column names}
}
\value{
None
}
\description{
Write tab-separate file, wraps \code{write.table}
}
