% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{clipboard}
\alias{clipboard}
\alias{clipboard}
\alias{pasteboard}
\title{Use OS clipboard}
\usage{
clipboard(x, delim = ",")

pasteboard(delim = "\\t", header = T)
}
\arguments{
\item{x}{Input object, can be piped through \code{margrittr} pipe}

\item{delim}{What is the delimiter of pasted object?}

\item{header}{Does pasted object have a header line?}
}
\value{
None
}
\description{
Copy/paste objects between R and system clipboard
}
