% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher-test.R
\name{fisher_test}
\alias{fisher_test}
\title{Fisher's test}
\usage{
fisher_test(a_positive, a_negative, b_positive, b_negative,
  alternative = "two.sided", conf_level = 0.95)
}
\arguments{
\item{alternative}{\code{two.sided} \code{greater} or \code{less}}

\item{a, a_total, b, b_total}{Test groups A and B}
}
\value{
Test statistics
}
\description{
Wraps \code{fisher.test} and tidies the output with \code{broom}
}
