% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher-test.R
\name{fisher_test}
\alias{fisher_test}
\title{Fisher's test}
\usage{
fisher_test(a_positive, a_negative, b_positive, b_negative,
  alternative = "two.sided", conf_level = 0.95)
}
\arguments{
\item{a_positive}{Upper left quadrant.}

\item{a_negative}{Upper right quadrant.}

\item{b_positive}{Lower left quadrant.}

\item{b_negative}{Lower right quadrant.}

\item{alternative}{\code{two.sided} \code{greater} or \code{less}}

\item{conf_level}{Level of confidence, if returning CIs.}
}
\value{
Test statistics
}
\description{
Wraps \code{fisher.test} and tidies the output with \code{broom}
}
