% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot-annotate-maf.R, R/oncokb-annotate-maf.R
\name{load_gene_annotation}
\alias{load_gene_annotation}
\alias{oncokb_annotate_maf}
\alias{query_oncokb}
\title{Annotate read MAF with OncoKB annotation}
\source{
\url{oncokb.org}

\url{github.com/oncokb/oncokb-annotator}
}
\usage{
load_gene_annotation()

query_oncokb(gene, protein_change, variant_type, start, end,
  cancer_type = "CANCER")

oncokb_annotate_maf(maf, cancer_types = NULL)
}
\arguments{
\item{maf}{Input MAF}

\item{cancer_types}{Data frame with samples mapped to cancer type for accurate levels of actionability}
}
\value{
Annotated MAF with columns indicating functionality of mutation and levels of actionability
}
\description{
Adds OncoKB oncogenicity and actionatility annotation to VEP-annotated MAF. See URLs below.
}
