% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_maf.R
\docType{data}
\name{test_maf}
\alias{test_maf}
\title{Example MAF for testing}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13 rows and 47 columns.}
\source{
\url{https://github.com/oncokb/oncokb-annotator}
}
\usage{
test_maf
}
\description{
Example MAF for testing
}
\keyword{datasets}
