% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/theme_bwmin.R
\name{theme_bwmin}
\alias{theme_bwmin}
\title{Black-and-white minimalistic theme for ggplot2}
\usage{
theme_bwmin(base_size = 12, base_family = "Helvetica")
}
\description{
Use by adding \code{+theme_bwmin()} to your ggplot call.
}

